Chunk loading of messages from file, removing forceTimeout
